[2022-06-23] Added modeline to 2022q2-season.md
Modelines give hints to text editors about how the file should be
displayed: tab size, whether to indent with tabs or spaces, the
syntax (if it cannot be guessed by the file extension), and so on.

The modeline added here is for kyoudai's preferences, and is a
type that should be supported by Sublime Text 2/3/4 (with the package
"Emacs-like Modelines" installed), as well as any other editor that
supports Emacs-style modelines.

Most of the preferences set in the modeline -- all of them that begin
with "st-", are Sublime Text specific, and should not interfere with
VSC, Emacs, or any other programmer's editor.

If you wish to use modelines, you must install an extension for VSC.
You can download that extension from https://bit.ly/3ncUJGk.