[2022q3] Add seasonal listing (alpha sort)
It took a while, but here it is! The matrix for the Summer 2022 season.
Please take a look and fill in your columns.

NOTE
====
When you have added your status and sent a PR, I will create
a second version sorted by start date, so please watch for that to
come through.

~kyoudai