[2022q3] Make undecided symbol more clear
I had set the symbol for 'undecided' to '☐', but then I realized that
it looks the same as the glyph that is shown when a character is
missing from the font in use. To avoid that, I changed it to '⁇',

I tried to use only glyphs/symbols that are included in nearly all
standard fonts, but if there is by chance something that doesn't
display correctly for you, then we can change it.

The reason I went with the symbols I did is because I wanted to
make it easier to search and replace. Something like 'W' and
'+' and '-' could also match titles and/or other parts of the
file that were not desired. By using these symbols that are
unique to the status columns, we can do a "select all" and be sure
that it will select _only_ the intended symbols.

~kyoudai, Mon Jun 27 16:47:15 PDT 2022